Add abyssal whip special attack
