+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/energy/Energy.kt
@@ -12,6 +12,6 @@ fun Player.energyPercent() = (runEnergy / MAX_ENERGY.toDouble() * 100).toInt()
 var Player.runEnergy: Int
     get() = this["energy", MAX_ENERGY]
     set(value) {
-        this["energy", true] = value
+        this["energy", true] = value.coerceIn(0, MAX_ENERGY)
         sendRunEnergy(energyPercent())
     }
